refactor(transaction-breakdown): transaction breakdown separated into contract and deploy components
