+++ src/components/bakoLoading/index.tsx
@@ -15,7 +15,7 @@ const defaultOptions = {
 const BakoLoading = () => {
   return (
     <Flex alignItems="center" justifyContent="center" w="full" h="100vh">
-      <Lottie {...defaultOptions} size={20} style={{ height: 180 }} />
+      <Lottie {...defaultOptions} style={{ height: 180 }} />
     </Flex>
   );
 };
