+++ src/modules/transactions/components/TransactionCard/deposit-details/index.ts
@@ -1,0 +1,1 @@
+export * from './DepositDetails';
