+++ src/modules/transactions/components/TransactionCard/transfer-details/ContractInfos.tsx
@@ -1,0 +1,92 @@
+import { MinimalAlertIcon } from '@/components';
+import {
+  Avatar,
+  Card,
+  Divider,
+  HStack,
+  Icon,
+  Text,
+  VStack,
+} from '@chakra-ui/react';
+import { TransactionUI } from '../Details';
+import { useQueryParams } from '@/modules';
+
+export interface IContractInfos {
+  transaction: TransactionUI;
+  isPending: boolean;
+  isNotSigned: boolean;
+}
+
+const ContractInfos = ({
+  transaction,
+  isPending,
+  isNotSigned,
+}: IContractInfos) => {
+  const { origin } = useQueryParams();
+  return (
+    <Card
+      bgColor="grey.825"
+      borderColor="grey.925"
+      borderRadius={10}
+      w={{ base: 'full', xs: 'unset' }}
+      px={5}
+      py={{ base: 2, xs: 4 }}
+      borderWidth="1px"
+      mt={4}
+    >
+      <Text color="grey.550" fontSize="xs">
+        Requesting a transaction from:
+      </Text>
+
+      <Divider borderColor="grey.950" my={4} />
+
+      <HStack width="100%" alignItems="center" spacing={4} h="32px">
+        <Avatar
+          borderRadius="6.4px"
+          color="white"
+          bgColor="dark.950"
+          name={transaction.name}
+          size="sm"
+        />
+        <VStack alignItems="flex-start" spacing={0}>
+          <Text variant="subtitle" fontSize="14px" color="grey.250">
+            {transaction.summary?.type}
+          </Text>
+          <Text color="brand.500" variant="description" fontSize="xs">
+            {origin?.split('//')[1]}
+            {/* bakoconnector-git-gr-featbakosafe-infinity-base.vercel.app */}
+          </Text>
+        </VStack>
+      </HStack>
+      {isPending && isNotSigned && (
+        <HStack
+          bg="warning.700"
+          borderColor="warning.700"
+          borderWidth="1px"
+          borderRadius={10}
+          mt={{ base: 4, xs: 8 }}
+          py={4}
+          px={4}
+        >
+          <Icon
+            as={MinimalAlertIcon}
+            color="warning.600"
+            fontSize={28}
+            mt={-7}
+          />
+
+          <VStack spacing={0} alignItems="flex-start">
+            <Text fontWeight="bold" color="#FFC010" fontSize="sm">
+              Double check it!
+            </Text>
+            <Text color="#EED07C" fontSize="xs">
+              Please carefully review this externally created transaction before
+              approving it.
+            </Text>
+          </VStack>
+        </HStack>
+      )}
+    </Card>
+  );
+};
+export { ContractInfos };
