+++ src/modules/transactions/components/TransactionCard/transfer-details/DeploymentInfos.tsx
