+++ src/modules/transactions/components/TransactionCard/transfer-details/index.tsx
@@ -1,0 +1,1 @@
+export * from './TransactionBreakdown';
