+++ src/modules/workspace/components/wkHomeTransactions/index.tsx
@@ -7,13 +7,13 @@ import {
   Spacer,
   Text,
 } from '@chakra-ui/react';
-import { css, keyframes } from '@emotion/react';
+import { css } from '@emotion/react';
 import { useEffect, useState } from 'react';
 import { MdKeyboardArrowRight } from 'react-icons/md';
 
 import { CustomSkeleton } from '@/components';
 import { EmptyState } from '@/components/emptyState';
-import { Pages, useScreenSize } from '@/modules/core';
+import { Pages, shakeAnimationX, useScreenSize } from '@/modules/core';
 import {
   TransactionCard,
   TransactionCardMobile,
@@ -23,14 +23,6 @@ import {
 import { useTransactionsContext } from '@/modules/transactions/providers/TransactionsProvider';
 
 import { useWorkspaceContext } from '../../WorkspaceProvider';
-
-const shakeAnimation = keyframes`
-  0% { transform: translateX(0); }
-  25% { transform: translateX(-2px); }
-  50% { transform: translateX(2px); }
-  75% { transform: translateX(-2px); }
-  100% { transform: translateX(0); }
-`;
 
 const WkHomeTransactions = () => {
   const [hasTransactions, setHasTransactions] = useState(false);
@@ -106,7 +98,7 @@ const WkHomeTransactions = () => {
           }
           css={css`
             &:hover .btn-icon {
-              animation: ${shakeAnimation} 0.5s ease-in-out;
+              animation: ${shakeAnimationX} 0.5s ease-in-out;
             }
           `}
           px={isExtraSmall ? 3 : 4}
